camera: allow loading files with text relocations from /system
the qcom camera driver needs to load a shared library from /system
which contains a text relocation. Allow it.

Due to bug 20013628, SELinux policy was inappropriately treating
an execmod denial as an execmem denial. Move to using a proper
execmod denial and get rid of execmem.

Addresses the following denial:
  avc: denied { execmod } for pid=208 comm="mm-qcamera-daem" path="/system/vendor/lib/libmmcamera_faceproc.so" dev="mmcblk0p22" ino=1739 scontext=u:r:camera:s0 tcontext=u:object_r:system_file:s0 tclass=file

Bug: 20013628
Change-Id: I9a1079b5e95390f1aebeeaeceaa0271f58c6b2de